#frontend/supply-page: рефакторинг кода
